Get zeta1's tests running
